requirements.txt
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
python-magic==0.4.13
